LPS-57181 move to module
